medicare_billing_pro/requirements.txt
streamlit>=1.28.0
supabase>=2.0.0
fpdf2>=2.7.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
